Add cookie function
Add the ability to pull the cookie from an external cookie.txt file in the same folder as the jar itself.  The cookie.txt file only has the value of fakku_sid and will import it directly to a variable.  This allows for access to information that only shows when you are logged in like controversial content.  Related to https://github.com/jobobby04/LanraragiMissingGalleries/issues/1